* ../common-src/amutil.c: fix sscanf portability without '%jd'.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@5178 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/amutil.c
@@ -1811,9 +1811,9 @@ parse_crc(
     crc_t *crc)
 {
     unsigned int c;
-    intmax_t size;
+    long long size;
 
-    if (sscanf(s, "%x:%jd", &c, &size) == 2) {
+    if (sscanf(s, "%x:%lld", &c, &size) == 2) {
         crc->crc = c;
         crc->size = size;
     }
